requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
transformers
pillow
requests
Jinja2
accelerate